refactor: update import paths and enhance SignUp component with error handling and user feedback
