+++ frontend/src/App.tsx
@@ -7,14 +7,17 @@ import { Navigate, Outlet, Route, Routes, useLocation } from "react-router-dom";
 import SearchProfile from "./pages/search-profile/SearchProfile";
 import Settings from "./pages/settings/Settings";	
 import UserProfile from "./pages/user-profile/UserProfile";
-import SignUp from "./pages/sign-up/SignUp";
-import { useUserAuth } from "./context/UserAuthContext";
-import Login from "./pages/login/Login";
+import SignUp from "./pages/signup-login/SignUp";
+import { useUserAuth } from "./context/userauth.context";
+import Login from "./pages/signup-login/Login";
 import nrLogo from "./assets/ng_logo.png"
 import UploadNote from "./pages/upload-note/UploadNote";
 import NotFound from "./pages/error-pages/NotFound";
-import { useGlobalComponentController } from "./context/GlobalComponentContext";
+import { useGlobalComponentController } from "./context/globaldata.context";
 import { CustomToast } from "./partials/Toast";
+import ProfessionSelection from "./pages/signup-login/ProfessionSelection";
+import Checkout from "./pages/checkout/SelectPlan";
+import Payment from "./pages/checkout/Payment";
 
 //TODO: A reddit like logo when the feed loads or the user auth loads
 
@@ -72,25 +75,26 @@ function App() {
 	const { toast: [toast, setToast] } = useGlobalComponentController()
 
 	return (
-		<Routes>
-			<Route element={<PublicRoute />}>
-				<Route path="/login" element={<Login />} />
-				<Route path="/signup" element={<SignUp />} />
-				<Route path="/profession" element={<ProfessionSelection/>} />
-				<Route path="/checkout" element={<Checkout/>} />
-				<Route path="/payment" element={<Payment/>} />
-			</Route>
-
-			<Route element={<ProtectedRoute />}>
-				<Route element={<MainLayout userAuth={userAuth}/>}>
-					<Route path="/" element={<DashBoard />} />
-					<Route path="/post/:postID" element={<PostView />} />
-					<Route path="/user/:username" element={<UserProfile />} />
-					<Route path="/search-profile" element={<SearchProfile />} />
-					<Route path="/settings" element={<Settings />} />
-					<Route path="/upload" element={<UploadNote />} />
+		<>
+			<Routes>
+				<Route element={<PublicRoute />}>
+					<Route path="/login" element={<Login />} />
+					<Route path="/signup" element={<SignUp />} />
+					<Route path="/profession" element={<ProfessionSelection/>} />
+					<Route path="/checkout" element={<Checkout/> } />
+					<Route path="/payment" element={<Payment/>} />
 				</Route>
-			</Route>
+
+				<Route element={<ProtectedRoute />}>
+					<Route element={<MainLayout userAuth={userAuth}/>}>
+						<Route path="/" element={<DashBoard />} />
+						<Route path="/post/:postID" element={<PostView />} />
+						<Route path="/user/:username" element={<UserProfile />} />
+						<Route path="/search-profile" element={<SearchProfile />} />
+						<Route path="/settings" element={<Settings />} />
+						<Route path="/upload" element={<UploadNote />} />
+					</Route>
+				</Route>
 
 				<Route path="*" element={<Navigate to="/not-found" state={{ type: "page", route: pathname }} replace={true} />} />
 				<Route path="/not-found" element={<NotFound />} />
